+++ src/patterns/index.ts
@@ -1,2 +1,10 @@
 export * from './chainOfResponsibility/middlewareDecorator'
 export * from './chainOfResponsibility/handlers'
+
+// usando esta referencia: https://doeken.org/blog/repository-pattern
+export * from './repository/assetRepository'
+export * from './repository/marketRepository'
+export * from './repository/orderRepository'
+export * from './repository/portfolioRepository'
+export * from './repository/transactionRepository'
+export * from './repository/userRepository'