+++ src/patterns/repository/portfolioRepository.ts
@@ -1,0 +1,19 @@
+import { Portfolio } from "../../models/portfolio";
+
+interface IPortfolioRepository {
+    getByUserId(userId: string): Portfolio | undefined
+    update(portfolio: Portfolio): void
+}
+
+export class PortfolioRepository implements IPortfolioRepository {
+    constructor(private data: Map<string, Portfolio> = new Map()) { }
+
+    getByUserId(userId: string): Portfolio | undefined {
+        return this.data.get(userId);
+    }
+
+    update(portfolio: Portfolio): void {
+        this.data.set(portfolio.userId, portfolio);
+    }
+}
+